Frontend move updating the FilterQuery from Filter.tsx to FilterEntityType.tsx
